Refactor version to v20 in all files #21
