+++ metadata/latest/build_source/schemas/context.json
@@ -1,6 +1,6 @@
 {
     "$schema": "http://json-schema.org/draft-07/schema#",
-    "$id": "general.json#metadata/v200_draft/schemas/context.json",
+    "$id": "general.json#metadata/v20/schemas/context.json",
     "type": "object",
     "properties": {
         "context": {
