+++ metadata/latest/build_source/schemas/contribution.json
@@ -1,6 +1,6 @@
 {
     "$schema": "http://json-schema.org/draft-07/schema#",
-    "$id": "general.json#metadata/v200_draft/schemas/contribution.json",
+    "$id": "general.json#metadata/v20/schemas/contribution.json",
     "type": "object",
     "properties": {
         "contributors": {
