+++ metadata/latest/build_source/schemas/general.json
@@ -1,6 +1,6 @@
 {
     "$schema": "http://json-schema.org/draft-07/schema#",
-    "$id": "general.json#metadata/v200_draft/schemas/general.json",
+    "$id": "general.json#metadata/v20/schemas/general.json",
     "type": "object",
     "properties": {
         "name": {
