+++ metadata/latest/build_source/schemas/json_ld.json
@@ -1,6 +1,6 @@
 {
     "$schema": "http://json-schema.org/draft-07/schema#",
-    "$id": "json_ld.json#metadata/v200_draft/schemas/json_ld.json",
+    "$id": "json_ld.json#metadata/v20/schemas/json_ld.json",
     "type": "object",
     "properties": {
         "@id": {
