+++ metadata/latest/build_source/schemas/licences.json
@@ -1,6 +1,6 @@
 {
     "$schema": "http://json-schema.org/draft-07/schema#",
-    "$id": "general.json#metadata/v200_draft/schemas/licences.json",
+    "$id": "general.json#metadata/v20/schemas/licences.json",
     "type": "object",
     "properties": {
         "licenses": {
