+++ metadata/latest/build_source/schemas/resources.json
@@ -1,6 +1,6 @@
 {
     "$schema": "http://json-schema.org/draft-07/schema#",
-    "$id": "general.json#metadata/v200_draft/schemas/resources.json",
+    "$id": "general.json#metadata/v20/schemas/resources.json",
     "type": "object",
     "properties": {
         "resources": {
