+++ metadata/latest/build_source/schemas/review.json
@@ -1,6 +1,6 @@
 {
     "$schema": "http://json-schema.org/draft-07/schema#",
-    "$id": "general.json#metadata/v200_draft/schemas/review.json",
+    "$id": "general.json#metadata/v20/schemas/review.json",
     "type": "object",
     "properties": {
         "review": {
