+++ metadata/latest/build_source/schemas/sources.json
@@ -1,6 +1,6 @@
 {
     "$schema": "http://json-schema.org/draft-07/schema#",
-    "$id": "general.json#metadata/v200_draft/schemas/sources.json",
+    "$id": "general.json#metadata/v20/schemas/sources.json",
     "type": "object",
     "properties": {
         "sources": {
