+++ metadata/latest/build_source/schemas/spatial.json
@@ -1,6 +1,6 @@
 {
     "$schema": "http://json-schema.org/draft-07/schema#",
-    "$id": "general.json#metadata/v200_draft/schemas/spatial.json",
+    "$id": "general.json#metadata/v20/schemas/spatial.json",
     "type": "object",
     "properties": {
         "spatial": {
