+++ metadata/latest/build_source/schemas/temporal.json
@@ -1,6 +1,6 @@
 {
     "$schema": "http://json-schema.org/draft-07/schema#",
-    "$id": "general.json#metadata/v200_draft/schemas/temporal.json",
+    "$id": "general.json#metadata/v20/schemas/temporal.json",
     "type": "object",
     "properties": {
         "temporal": {
