+++ metadata/latest/build_source/scripts/resolve_schema_refs.py
@@ -36,7 +36,7 @@ logger = logging.getLogger(__name__)
 
 # Constants
 BASE_PATH = Path("metadata/")
-VERSION = "v200_draft"
+VERSION = "v20"
 VERSION_PATH = BASE_PATH / VERSION
 SCHEMA_BUILD_PATH = VERSION_PATH / "build"
 MAIN_SCHEMA_PATH = SCHEMA_BUILD_PATH / "main_schema.json"
