+++ metadata/v20/build_source/schema_structure.json
@@ -1,11 +1,11 @@
 {
     "$schema": "http://json-schema.org/draft-07/schema#",
     "$id": "https://raw.githubusercontent.com/OpenEnergyPlatform/oemetadata/master/oemetadata/latest/schema.json",
-    "description": "Open Energy Plaftorm (OEP) metadata schema latest",
+    "description": "Open Energy Platform (OEP) metadata schema latest",
     "type": "object",
     "properties": {
-        "generalCollection": {
-            "$ref": "general_collection.json#"
+        "collection": {
+            "$ref": "collection.json#"
         },
         "resources": {
             "description": "A collection of related data resource descriptions and metadata. This may include relational data models which are technically related",
