+++ metadata/v20/build_source/schemas/collection.json
@@ -1,6 +1,6 @@
 {
     "$schema": "http://json-schema.org/draft-07/schema#",
-    "$id": "general.json#metadata/v200_draft/schemas/general_collection.json",
+    "$id": "general.json#metadata/v20/schemas/collection.json",
     "type": "object",
     "properties": {
         "name": {
