+++ metadata/v20/build_source/schemas/meta.json
@@ -1,6 +1,6 @@
 {
     "$schema": "http://json-schema.org/draft-07/schema#",
-    "$id": "general.json#metadata/v200_draft/schemas/meta.json",
+    "$id": "general.json#metadata/v20/schemas/meta.json",
     "type": "object",
     "properties": {
         "metaMetadata": {
