+++ metadata/v20/build_source/schemas/provenance.json
@@ -1,6 +1,6 @@
 {
     "$schema": "http://json-schema.org/draft-07/schema#",
-    "$id": "general.json#metadata/v200_draft/schemas/provenance.json",
+    "$id": "general.json#metadata/v20/schemas/provenance.json",
     "type": "object",
     "properties": {
         "provenance": {
