+++ metadata/v20/build_source/schemas/resource.json
@@ -1,6 +1,6 @@
 {
     "$schema": "http://json-schema.org/draft-07/schema#",
-    "$id": "general.json#metadata/v200_draft/schemas/resource.json",
+    "$id": "general.json#metadata/v20/schemas/resource.json",
     "type": "object",
     "properties": {
             "type": {
