+++ tests/metadata/v20/test_schema.py
@@ -1,12 +1,12 @@
 def test_if_schema_json_loads_successfully():
     try:
-        from metadata.v200_draft.schema import OEMETADATA_V200_SCHEMA
+        from metadata.v20.schema import OEMETADATA_V200_SCHEMA
     except Warning:
         print("Metadata Schema v2.0.0 cant load. Check if the files are missing!")
 
 
 def test_if_schema_json_has_correct_schema_and_id_set():
-    from metadata.v200_draft.schema import OEMETADATA_V200_SCHEMA
+    from metadata.v20.schema import OEMETADATA_V200_SCHEMA
     import string
 
     def get_string(s):
@@ -23,7 +23,7 @@ def test_if_schema_json_has_correct_schema_and_id_set():
 
 def test_schema_against_metaschema_which_should_succeed():
     import jsonschema
-    from metadata.v200_draft.schema import OEMETADATA_V200_SCHEMA
+    from metadata.v20.schema import OEMETADATA_V200_SCHEMA
     from metadata.metaschema.draft07.schema import OEMETADATA_METASCHEMA_DRAFT07_SCHEMA
 
     assert (
